Setting up the Express Server
Our server will accept an HTTP request on port 3000. It creates an Async Hook to track every request by calling the createRequestContext() in a middleware function - a function that has access to an HTTP's request and response objects. The server then sends a JSON response with the data captured by the Async Hook.